+++ PyWinRT/TypeExtensions.cs
@@ -338,6 +338,9 @@ static class TypeExtensions
                 when gen.ElementType.FullName == "Windows.Foundation.IReference`1"
                 => $"typing.Optional[{string.Join(", ", gen.GenericArguments.Select(p => p.ToPyTypeName(ns, map, quoteImportedTypes)))}]",
             GenericInstanceType gen
+                when gen.ElementType.FullName == "Windows.Foundation.Collections.IVector`1"
+                => $"typing.MutableSequence[{gen.GenericArguments[0].ToPyTypeName(ns, map, quoteImportedTypes)}]",
+            GenericInstanceType gen
                 => $"{(gen.Namespace == ns ? "" : $"{(quoteImportedTypes ? "\"" : "")}{gen.Namespace.ToPyModuleAlias()}.")}{gen.Name.ToNonGeneric()}[{string.Join(", ", gen.GenericArguments.Select(p => p.ToPyTypeName(ns, map)))}]{(gen.Namespace != ns && quoteImportedTypes ? "\"" : "")}",
             ByReferenceType t => t.ElementType.ToPyTypeName(ns, map, quoteImportedTypes),
             OptionalModifierType t => t.ElementType.ToPyTypeName(ns, map, quoteImportedTypes),